+++ pkg/network/ebpf/c/sk/sk_tcp.h
@@ -10,299 +10,12 @@
 #include "tracer/tracer.h"
 #include "tracer/port.h"
 #include "sk.h"
+#include "sk_conn.h"
 
 __maybe_unused static __always_inline bool tcp_failed_connections_enabled() {
     __u64 val = 0;
     LOAD_CONSTANT("tcp_failed_connections_enabled", val);
     return val > 0;
-}
-
-static __always_inline int create_tcp_conn(conn_t *conn, struct sock *sk, sk_tcp_stats_t *sk_stats, struct task_struct *task) {
-    struct tcp_sock *tp = bpf_skc_to_tcp_sock(sk);
-    if (!tp) {
-        return 0;
-    }
-
-    if (sk_stats) {
-        copy_conn_tuple(&conn->tup, &sk_stats->tup);
-    }
-    if (!read_conn_tuple_sk(&conn->tup, sk, task)) {
-        return 0;
-    }
-    conn->tup.metadata |= CONN_TYPE_TCP;
-
-    conn->tcp_stats.rtt = tp->srtt_us >> 3;
-    conn->tcp_stats.rtt_var = tp->mdev_us >> 2;
-    conn->tcp_stats.retransmits = tp->total_retrans;
-    conn->tcp_stats.reord_seen = tp->reord_seen;
-    conn->tcp_stats.rcv_ooopack = tp->rcv_ooopack;
-    conn->tcp_stats.delivered_ce = tp->delivered_ce;
-    conn->tcp_stats.ecn_negotiated = tp->ecn_flags & TCP_ECN_OK;
-
-    conn->conn_stats.sent_bytes = tp->bytes_sent;
-    conn->conn_stats.recv_bytes = tp->bytes_received;
-    conn->conn_stats.sent_packets = tp->segs_out;
-    conn->conn_stats.recv_packets = tp->segs_in;
-    conn->conn_stats.timestamp = tp->tcp_mstamp * NSEC_PER_USEC;
-    conn->conn_stats.cookie = (__u32)(bpf_get_socket_cookie(sk) & 0xFFFFFFFF);
-    // TODO conn->conn_stats.protocol_stack
-    // TODO conn->conn_stats.tls_tags
-    // TODO conn->conn_stats.cert_id
-
-    if (sk_stats) {
-        conn->tcp_stats.failure_reason = sk_stats->failure_reason;
-        conn->tcp_stats.state_transitions = sk_stats->state_transitions;
-        conn->tcp_stats.tcp_event_stats = sk_stats->tcp_event_stats;
-        conn->conn_stats.duration_ms = sk_stats->start_ms;
-        conn->conn_stats.direction = sk_stats->direction;
-
-        // offset absolute counters with initial values read
-        conn->conn_stats.sent_bytes -= sk_stats->initial_sent_bytes;
-        conn->conn_stats.recv_bytes -= sk_stats->initial_recv_bytes;
-        conn->conn_stats.sent_packets -= sk_stats->initial_sent_packets;
-        conn->conn_stats.recv_packets -= sk_stats->initial_recv_packets;
-        conn->tcp_stats.retransmits -= sk_stats->initial_retransmits;
-    }
-    return 1;
-}
-
-static __always_inline int create_udp_conn(conn_t *conn, struct sock *sk, sk_udp_stats_t *sk_stats, struct task_struct *task) {
-    if (sk_stats) {
-        copy_conn_tuple(&conn->tup, &sk_stats->tup);
-    }
-    if (!read_conn_tuple_sk(&conn->tup, sk, task)) {
-        return 0;
-    }
-    conn->tup.metadata |= CONN_TYPE_UDP;
-
-//    struct net *nt = sk->sk_net.net;
-//    conn->conn_stats.sent_packets = nt->mib.udp_statistics->mibs[UDP_MIB_OUTDATAGRAMS];
-//    conn->conn_stats.recv_packets = nt->mib.udp_statistics->mibs[UDP_MIB_INDATAGRAMS];
-
-    conn->conn_stats.cookie = (__u32)(bpf_get_socket_cookie(sk) & 0xFFFFFFFF);
-    // TODO conn->conn_stats.protocol_stack
-    // TODO conn->conn_stats.tls_tags
-    // TODO conn->conn_stats.cert_id
-
-    if (sk_stats) {
-        conn->conn_stats.duration = sk_stats->start_ns;
-        conn->conn_stats.direction = sk_stats->direction;
-        conn->conn_stats.sent_bytes = sk_stats->sent_bytes;
-        conn->conn_stats.recv_bytes = sk_stats->recv_bytes;
-        conn->conn_stats.sent_packets = sk_stats->sent_packets;
-        conn->conn_stats.recv_packets = sk_stats->recv_packets;
-        conn->conn_stats.timestamp = sk_stats->timestamp_ns;
-        conn->conn_stats.flags = sk_stats->flags;
-    }
-    return 1;
-}
-
-static __always_inline void initialize_tcp_socket(struct sock *sk, struct task_struct *task) {
-    sk_tcp_stats_t *sk_stats = bpf_sk_storage_get(&sk_tcp_stats, sk, 0, BPF_SK_STORAGE_GET_F_CREATE);
-    if (!sk_stats) {
-        return;
-    }
-    struct tcp_sock *tp = bpf_skc_to_tcp_sock(sk);
-    if (!tp) {
-        return;
-    }
-
-    // TODO do any of these need to be stored?
-//    conn->tcp_stats.reord_seen = tp->reord_seen;
-//    conn->tcp_stats.rcv_ooopack = tp->rcv_ooopack;
-//    conn->tcp_stats.delivered_ce = tp->delivered_ce;
-
-    sk_stats->initial_sent_bytes = tp->bytes_sent;
-    sk_stats->initial_recv_bytes = tp->bytes_received;
-    sk_stats->initial_sent_packets = tp->segs_out;
-    sk_stats->initial_recv_packets = tp->segs_in;
-    sk_stats->initial_retransmits = tp->total_retrans;
-
-    sk_stats->tup.pid = task->tgid;
-    sk_stats->tup.netns = get_netns_from_sock(sk);
-
-    port_binding_t pb = {};
-    pb.netns = sk_stats->tup.netns;
-    pb.port = read_sport(sk);
-    u32 *port_count = bpf_map_lookup_elem(&port_bindings, &pb);
-    sk_stats->direction = (port_count != NULL && *port_count > 0) ? CONN_DIRECTION_INCOMING : CONN_DIRECTION_OUTGOING;
-}
-
-static __always_inline void initialize_udp_socket(struct sock *sk, struct task_struct *task) {
-    sk_udp_stats_t *sk_stats = bpf_sk_storage_get(&sk_udp_stats, sk, 0, BPF_SK_STORAGE_GET_F_CREATE);
-    if (!sk_stats) {
-        return;
-    }
-
-    sk_stats->tup.pid = task->tgid;
-    sk_stats->tup.netns = get_netns_from_sock(sk);
-
-    port_binding_t pb = {};
-    pb.netns = sk_stats->tup.netns;
-    pb.port = read_sport(sk);
-    u32 *port_count = bpf_map_lookup_elem(&udp_port_bindings, &pb);
-    sk_stats->direction = (port_count != NULL && *port_count > 0) ? CONN_DIRECTION_INCOMING : CONN_DIRECTION_OUTGOING;
-}
-
-SEC("iter/task_file")
-int bpf_iter__task_file_port_bindings(struct bpf_iter__task_file *ctx) {
-    struct task_struct *task = ctx->task;
-    struct file *file = ctx->file;
-    if (!task || !file) {
-        return 0;
-    }
-    struct socket *sock = bpf_sock_from_file(file);
-    if (!sock) {
-        return 0;
-    }
-    struct sock *sk = sock->sk;
-
-    if (sk->sk_protocol == IPPROTO_TCP || sk->sk_protocol == IPPROTO_MPTCP) {
-        switch (sk->sk_family) {
-        case AF_INET6:
-            if (!is_tcpv6_enabled()) return 0;
-            break;
-        case AF_INET:
-            if (!is_tcpv4_enabled()) return 0;
-            break;
-        default:
-            return 0;
-        }
-
-        if (sk->__sk_common.skc_state == TCP_LISTEN) {
-            port_binding_t pb = {};
-            pb.netns = get_netns_from_sock(sk);
-            pb.port = read_sport(sk);
-            add_port_bind(&pb, port_bindings);
-        }
-        return 0;
-    } else if (sk->sk_protocol == IPPROTO_UDP) {
-        switch (sk->sk_family) {
-        case AF_INET6:
-            if (!is_udpv6_enabled()) return 0;
-            break;
-        case AF_INET:
-            if (!is_udpv4_enabled()) return 0;
-            break;
-        default:
-            return 0;
-        }
-
-        // TODO skip if in ephemeral port range
-        if (sk->__sk_common.skc_state == TCP_CLOSE) {
-            port_binding_t pb = {};
-            pb.netns = get_netns_from_sock(sk);
-            pb.port = read_sport(sk);
-            add_port_bind(&pb, udp_port_bindings);
-        }
-        return 0;
-    }
-    return 0;
-}
-
-SEC("iter/task_file")
-int bpf_iter__task_file_initial_sockets(struct bpf_iter__task_file *ctx) {
-    struct task_struct *task = ctx->task;
-    struct file *file = ctx->file;
-    if (!task || !file) {
-        return 0;
-    }
-    struct socket *sock = bpf_sock_from_file(file);
-    if (!sock) {
-        return 0;
-    }
-    struct sock *sk = sock->sk;
-
-    if (sk->sk_protocol == IPPROTO_TCP || sk->sk_protocol == IPPROTO_MPTCP) {
-        switch (sk->sk_family) {
-        case AF_INET6:
-            if (!is_tcpv6_enabled()) return 0;
-            break;
-        case AF_INET:
-            if (!is_tcpv4_enabled()) return 0;
-            break;
-        default:
-            return 0;
-        }
-
-        initialize_tcp_socket(sk, task);
-        return 0;
-    } else if (sk->sk_protocol == IPPROTO_UDP) {
-        switch (sk->sk_family) {
-        case AF_INET6:
-            if (!is_udpv6_enabled()) return 0;
-            break;
-        case AF_INET:
-            if (!is_udpv4_enabled()) return 0;
-            break;
-        default:
-            return 0;
-        }
-
-        initialize_udp_socket(sk, task);
-        return 0;
-    }
-    return 0;
-}
-
-SEC("iter/task_file")
-int bpf_iter__task_file_socket(struct bpf_iter__task_file *ctx) {
-    struct task_struct *task = ctx->task;
-    struct file *file = ctx->file;
-    struct seq_file *seq = ctx->meta->seq;
-    if (!task || !file) {
-        return 0;
-    }
-    struct socket *sock = bpf_sock_from_file(file);
-    if (!sock) {
-        return 0;
-    }
-    struct sock *sk = sock->sk;
-
-    conn_t conn = {};
-    if (sk->sk_protocol == IPPROTO_TCP || sk->sk_protocol == IPPROTO_MPTCP) {
-        switch (sk->sk_family) {
-        case AF_INET6:
-            if (!is_tcpv6_enabled()) return 0;
-            break;
-        case AF_INET:
-            if (!is_tcpv4_enabled()) return 0;
-            break;
-        default:
-            return 0;
-        }
-
-        log_debug("iterate tcp: sk=%pK pid=%d", sk, task->tgid);
-        sk_tcp_stats_t *sk_stats = bpf_sk_storage_get(&sk_tcp_stats, sk, 0, 0);
-        if (!create_tcp_conn(&conn, sk, sk_stats, task)) {
-            return 0;
-        }
-    } else if (sk->sk_protocol == IPPROTO_UDP) {
-        switch (sk->sk_family) {
-        case AF_INET6:
-            if (!is_udpv6_enabled()) return 0;
-            break;
-        case AF_INET:
-            if (!is_udpv4_enabled()) return 0;
-            break;
-        default:
-            return 0;
-        }
-
-        log_debug("iterate udp: sk=%pK pid=%d", sk, task->tgid);
-        sk_udp_stats_t *sk_stats = bpf_sk_storage_get(&sk_udp_stats, sk, 0, 0);
-        if (!create_udp_conn(&conn, sk, sk_stats, task)) {
-            return 0;
-        }
-    } else {
-        return 0;
-    }
-
-    if (!conn.tup.pid) {
-        conn.tup.pid = task->tgid;
-    }
-    bpf_seq_write(seq, &conn, sizeof(conn_t));
-    return 0;
 }
 
 // get_tcp_failure returns an error code for tcp_done/tcp_close, if there was one
@@ -375,8 +88,10 @@ int tcp_sockops(struct bpf_sock_ops *ctx) {
         bpf_sock_ops_cb_flags_set(ctx, BPF_SOCK_OPS_STATE_CB_FLAG);
         return 1;
     case BPF_SOCK_OPS_STATE_CB:
+        break;
     case BPF_SOCK_OPS_ACTIVE_ESTABLISHED_CB:
     case BPF_SOCK_OPS_PASSIVE_ESTABLISHED_CB:
+        bpf_sock_ops_cb_flags_set(ctx, BPF_SOCK_OPS_STATE_CB_FLAG);
         break;
     default:
         return 1;
